web/requirements.txt
flask==3.0.2
paho-mqtt==2.1.0